collaboration/requirements.txt
asgiref==3.8.1
distlib==0.3.9
Django==4.2.16
django-debug-toolbar==4.4.6
djangorestframework==3.15.2
drf-nested-routers==0.94.1
filelock==3.16.1
git-filter-repo==2.38.0
mysqlclient==2.2.6
pillow==11.0.0
platformdirs==4.3.6
sqlparse==0.5.2
tzdata==2024.2
virtualenv==20.27.1
